add liquidation table
