+++ containers/EmpSponsors.ts
@@ -12,7 +12,7 @@ import { getLiquidationPrice } from "../utils/getLiquidationPrice";
 import { isPricefeedInvertedFromTokenSymbol } from "../utils/getOffchainPrice";
 
 import { useQuery } from "@apollo/client";
-import { EMP_DATA } from "../apollo/uma/queries";
+import { ACTIVE_POSITIONS } from "../apollo/uma/queries";
 
 // Interfaces for dApp state storage.
 interface SponsorPositionState {
@@ -30,26 +30,9 @@ interface PositionQuery {
   collateral: string;
 }
 
-interface LiquidationQuery {
-  sponsor: { id: string };
-  liquidationId: string;
-  liquidator: { address: string };
-  disputer: { address: string };
-  tokensLiquidated: string;
-  lockedCollateral: string;
-  liquidatedCollateral: string;
-  status: string;
-  events: {
-    __typename: string;
-    tx_hash: string;
-    timestamp: string;
-  }[];
-}
-
 interface FinancialContractQuery {
   id: string;
   sponsorPositions: PositionQuery;
-  sponsorLiquidations: LiquidationQuery;
 }
 
 const useEmpSponsors = () => {
@@ -64,7 +47,7 @@ const useEmpSponsors = () => {
   // We set the poll interval to a very slow 5 seconds for now since the position states
   // are not expected to change much.
   // Source: https://www.apollographql.com/docs/react/data/queries/#polling
-  const { loading, error, data } = useQuery(EMP_DATA, {
+  const { loading, error, data } = useQuery(ACTIVE_POSITIONS, {
     context: { clientName: "UMA" },
     pollInterval: 5000,
   });
@@ -80,13 +63,11 @@ const useEmpSponsors = () => {
   };
 
   const [activePositions, setActivePositions] = useState<SponsorMap>({});
-  const [liquidations, setLiquidations] = useState<SponsorMap>({});
 
   // get position information about every sponsor that has ever created a position.
   const querySponsors = () => {
     // Start with a fresh table.
     setActivePositions({});
-    setLiquidations({});
 
     if (
       emp !== null &&
@@ -106,7 +87,6 @@ const useEmpSponsors = () => {
 
         if (empData) {
           let newPositions: SponsorMap = {};
-          let newLiquidations: SponsorMap = {};
 
           const collReqFromWei = parseFloat(
             fromWei(collateralRequirement, collDecs)
@@ -136,36 +116,7 @@ const useEmpSponsors = () => {
             };
           });
 
-          empData.liquidations.forEach((liquidation: LiquidationQuery) => {
-            const sponsor = utils.getAddress(liquidation.sponsor.id);
-            const liquidationCreatedEvent = liquidation.events.find(
-              (e) => e.__typename === "LiquidationCreatedEvent"
-            );
-            if (liquidationCreatedEvent) {
-              const liquidatedCR =
-                parseFloat(liquidation.tokensLiquidated) > 0
-                  ? parseFloat(liquidation.liquidatedCollateral) /
-                    parseFloat(liquidation.tokensLiquidated)
-                  : 0;
-              // There should always be a LiquidationCreatedEvent associated with each liquidation object, but if not then
-              // we will just ignore this strange edge case.
-              newLiquidations[sponsor] = {
-                sponsor,
-                liquidator: liquidation.liquidator?.address,
-                disputer: liquidation.disputer?.address,
-                liquidationId: liquidation.liquidationId,
-                liquidatedCR: liquidatedCR.toString(),
-                tokensLiquidated: liquidation.tokensLiquidated,
-                lockedCollateral: liquidation.lockedCollateral,
-                status: liquidation.status,
-                liquidationTimestamp: liquidationCreatedEvent.timestamp,
-                liquidationReceipt: liquidationCreatedEvent.tx_hash,
-              };
-            }
-          });
-
           setActivePositions(newPositions);
-          setLiquidations(newLiquidations);
         }
       }
     }
@@ -176,7 +127,7 @@ const useEmpSponsors = () => {
     querySponsors();
   }, [emp, data, latestPrice]);
 
-  return { activeSponsors: activePositions, liquidations: liquidations };
+  return { activeSponsors: activePositions };
 };
 
 const EmpSponsors = createContainer(useEmpSponsors);
